Update README.md
The users can set reminders for medicines that need to be taken, as well as for setting up the dates for upcoming appointments. the appication also has a geo-locator library that will help their family members keep a track of their location in case they are not able to reach them.
The application also consists of an SOS button. if the user(elderly person) is in such a situation, with a single click they can send an emergency notification to the family member and also inform the authorities.